requirements.txt
rapidfuzz
google-generativeai
flask
gunicorn
psycopg2-binary
